requirements.txt
flask
pandas
requests
python-dotenv

